+++ src/main/java/com/griddynamics/bigdata/CustomizableJob.java
@@ -6,7 +6,7 @@ import org.apache.commons.cli.ParseException;
 import org.apache.hadoop.conf.Configured;
 import org.apache.hadoop.fs.FileSystem;
 import org.apache.hadoop.fs.Path;
-import org.apache.hadoop.io.IntWritable;
+import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Job;
 import org.apache.hadoop.mapreduce.Mapper;
@@ -52,7 +52,7 @@ public abstract class CustomizableJob extends Configured implements Tool {
         job.setReducerClass(getReducerClass());
 
         job.setOutputKeyClass(Text.class);
-        job.setOutputValueClass(IntWritable.class);
+        job.setOutputValueClass(LongWritable.class);
 
         job.setInputFormatClass(getInputFormatClass());
